# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"IInstitution": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"institutionType": {
						"type": "string"
					},
					"level": {
						"type": "string"
					},
					"hasPharmacy": {
						"type": "boolean"
					},
					"admin": {
						"properties": {
							"email": {
								"type": "string"
							},
							"phone": {
								"type": "string"
							},
							"name": {
								"type": "string"
							}
						},
						"required": [
							"email",
							"phone",
							"name"
						],
						"type": "object"
					},
					"details": {
						"properties": {
							"percentage": {
								"type": "number",
								"format": "double"
							},
							"TIN": {
								"type": "string"
							},
							"location": {
								"type": "string"
							}
						},
						"required": [
							"percentage",
							"TIN",
							"location"
						],
						"type": "object"
					},
					"institutionId": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"institutionType",
					"level",
					"hasPharmacy",
					"admin",
					"details",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UserReponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"institutionId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"phone",
					"institutionId",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IServiceAct": {
				"properties": {
					"id": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"desc": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"serviceId": {
						"type": "string"
					},
					"institutionId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"service": {
						"$ref": "#/components/schemas/IService"
					},
					"institution": {
						"$ref": "#/components/schemas/IInstitution"
					}
				},
				"required": [
					"id",
					"label",
					"desc",
					"price",
					"serviceId",
					"institutionId",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IService": {
				"properties": {
					"id": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"desc": {
						"type": "string"
					},
					"level": {
						"type": "string"
					},
					"assignDuringOrientation": {
						"type": "boolean"
					},
					"institutionId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"acts": {
						"items": {
							"$ref": "#/components/schemas/IServiceAct"
						},
						"type": "array"
					},
					"institution": {
						"$ref": "#/components/schemas/IInstitution"
					}
				},
				"required": [
					"id",
					"label",
					"desc",
					"level",
					"assignDuringOrientation",
					"institutionId",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IServiceActRequest": {
				"properties": {
					"label": {
						"type": "string"
					},
					"institutionId": {
						"type": "string"
					},
					"desc": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"serviceId": {
						"type": "string"
					},
					"service": {
						"$ref": "#/components/schemas/IService"
					},
					"institution": {
						"$ref": "#/components/schemas/IInstitution"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"label",
					"institutionId",
					"desc",
					"price",
					"serviceId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IInstitutionDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"institutionType": {
						"type": "string"
					},
					"level": {
						"type": "string"
					},
					"hasPharmacy": {
						"type": "boolean"
					},
					"admin": {
						"properties": {
							"email": {
								"type": "string"
							},
							"phone": {
								"type": "string"
							},
							"name": {
								"type": "string"
							}
						},
						"required": [
							"email",
							"phone",
							"name"
						],
						"type": "object"
					},
					"details": {
						"properties": {
							"percentage": {
								"type": "number",
								"format": "double"
							},
							"TIN": {
								"type": "string"
							},
							"location": {
								"type": "string"
							}
						},
						"required": [
							"percentage",
							"TIN",
							"location"
						],
						"type": "object"
					},
					"institutionId": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"users": {
						"items": {
							"$ref": "#/components/schemas/UserReponse"
						},
						"type": "array"
					},
					"parentInstitution": {
						"$ref": "#/components/schemas/IInstitution"
					},
					"branches": {
						"items": {
							"$ref": "#/components/schemas/IInstitution"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"institutionType",
					"level",
					"hasPharmacy",
					"admin",
					"details",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IInstitutionResponse": {
				"properties": {
					"type": {
						"type": "string"
					},
					"rows": {
						"items": {
							"$ref": "#/components/schemas/IInstitutionDTO"
						},
						"type": "array"
					}
				},
				"required": [
					"type",
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPaged_IInstitutionResponse_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/IInstitutionResponse"
					},
					"totalItems": {
						"type": "number",
						"format": "double"
					},
					"currentPage": {
						"type": "number",
						"format": "double"
					},
					"itemsPerPage": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"totalItems",
					"currentPage",
					"itemsPerPage"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IInstitutionRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"institutionId": {
						"type": "string"
					},
					"level": {
						"type": "string"
					},
					"admin": {
						"properties": {
							"email": {
								"type": "string"
							},
							"phone": {
								"type": "string"
							},
							"name": {
								"type": "string"
							}
						},
						"required": [
							"email",
							"phone",
							"name"
						],
						"type": "object"
					},
					"details": {
						"properties": {
							"percentage": {
								"type": "number",
								"format": "double"
							},
							"TIN": {
								"type": "string"
							},
							"location": {
								"type": "string"
							}
						},
						"required": [
							"percentage",
							"TIN",
							"location"
						],
						"type": "object"
					},
					"institutionType": {
						"type": "string",
						"nullable": true
					},
					"hasPharmacy": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"level",
					"admin",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICreateBranch": {
				"properties": {
					"name": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"details": {
						"properties": {
							"percentage": {
								"type": "number",
								"format": "double"
							},
							"TIN": {
								"type": "string"
							},
							"location": {
								"type": "string"
							}
						},
						"required": [
							"percentage",
							"TIN",
							"location"
						],
						"type": "object"
					}
				},
				"required": [
					"name",
					"location",
					"details"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IInstitutionExam": {
				"properties": {
					"id": {
						"type": "string"
					},
					"institutionId": {
						"type": "string"
					},
					"examId": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"institutionId",
					"examId",
					"price",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IInsuranceExam": {
				"properties": {
					"id": {
						"type": "string"
					},
					"institutionId": {
						"type": "string"
					},
					"examId": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"institutionId",
					"examId",
					"price",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IExam": {
				"properties": {
					"id": {
						"type": "string"
					},
					"exam_code": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"institutionExam": {
						"items": {
							"$ref": "#/components/schemas/IInstitutionExam"
						},
						"type": "array"
					},
					"insuranceExam": {
						"items": {
							"$ref": "#/components/schemas/IInsuranceExam"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"exam_code",
					"description",
					"name",
					"price",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IInstitutionDrug": {
				"properties": {
					"id": {
						"type": "string"
					},
					"drugId": {
						"type": "string"
					},
					"institutionId": {
						"type": "string"
					},
					"purchaseId": {
						"type": "string"
					},
					"drugPurchaseId": {
						"type": "string"
					},
					"insuranceDrugId": {
						"type": "string",
						"nullable": true
					},
					"batchNumber": {
						"type": "string",
						"nullable": true
					},
					"expireDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"itemNo": {
						"type": "string"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"isAvailable": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"totalQuantity": {
						"type": "number",
						"format": "double"
					},
					"drug": {
						"$ref": "#/components/schemas/IDrug"
					},
					"insuranceDrug": {
						"$ref": "#/components/schemas/IInsuranceDrug"
					}
				},
				"required": [
					"id",
					"drugId",
					"institutionId",
					"purchaseId",
					"drugPurchaseId",
					"insuranceDrugId",
					"batchNumber",
					"expireDate",
					"itemNo",
					"quantity",
					"price",
					"isAvailable",
					"createdAt",
					"deletedAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IExamResponse": {
				"properties": {
					"rows": {
						"items": {
							"$ref": "#/components/schemas/IExam"
						},
						"type": "array"
					}
				},
				"required": [
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPaged_IExamResponse_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/IExamResponse"
					},
					"totalItems": {
						"type": "number",
						"format": "double"
					},
					"currentPage": {
						"type": "number",
						"format": "double"
					},
					"itemsPerPage": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"totalItems",
					"currentPage",
					"itemsPerPage"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IExamRequest": {
				"$ref": "#/components/schemas/Omit_IExam.id-or-createdAt-or-institutionExam_"
			},
...
			"IDrugDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"drug_code": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"designation": {
						"type": "string"
					},
					"instruction": {
						"type": "string",
						"nullable": true
					},
					"drugCategory": {
						"type": "string"
					},
					"isOnMarket": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"totalQuantity": {
						"type": "number",
						"format": "double"
					},
					"insuranceDrug": {
						"items": {
							"$ref": "#/components/schemas/IInsuranceDrug"
						},
						"type": "array"
					},
					"institution": {
						"$ref": "#/components/schemas/IInstitution"
					}
				},
				"required": [
					"id",
					"drug_code",
					"description",
					"designation",
					"instruction",
					"drugCategory",
					"isOnMarket",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IPaged_IDrugResponse_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/IDrugResponse"
					},
					"totalItems": {
						"type": "number",
						"format": "double"
					},
					"currentPage": {
						"type": "number",
						"format": "double"
					},
					"itemsPerPage": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"totalItems",
					"currentPage",
					"itemsPerPage"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IPaged_IInstitutionDrugResponse_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/IInstitutionDrugResponse"
					},
					"totalItems": {
						"type": "number",
						"format": "double"
					},
					"currentPage": {
						"type": "number",
						"format": "double"
					},
					"itemsPerPage": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"totalItems",
					"currentPage",
					"itemsPerPage"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IDrugRequest": {
				"properties": {
					"description": {
						"type": "string"
					},
					"designation": {
						"type": "string"
					},
					"drugCategory": {
						"type": "string"
					},
					"instruction": {
						"type": "string"
					},
					"drug_code": {
						"type": "string"
					},
					"totalQuantity": {
						"type": "number",
						"format": "double"
					},
					"insuranceDrug": {
						"items": {
							"$ref": "#/components/schemas/IInsuranceDrug"
						},
						"type": "array"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"drugId": {
						"type": "string"
					},
					"type": {
						"type": "string"
					}
				},
				"required": [
					"description",
					"designation",
					"drugCategory",
					"instruction",
					"drug_code"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwtAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Bearer token to access these API endpoints"
			}
		}
	},
	"info": {
		"title": "digital-med",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Reponse Ashimwe"
		}
	},
	"paths": {
...
		"/api/service-acts/{id}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IServiceAct"
								}
							}
						}
					}
				},
				"tags": [
					"Service Acts"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Service Acts"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IServiceActRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Service Acts"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/institutions": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaged_IInstitutionResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Institutions"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IInstitution"
								}
							}
						}
					}
				},
				"tags": [
					"Institutions"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IInstitutionRequest"
							}
						}
					}
				}
			}
		},
		"/api/institutions/{id}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IInstitutionDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Institutions"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Institutions"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IInstitutionRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Institutions"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/institutions/pharmacies": {
			"get": {
				"operationId": "GetPharmacies",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IInstitutionDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Institutions"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/institutions/insurances": {
			"get": {
				"operationId": "GetInsurances",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IInstitutionDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Institutions"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/institutions/branches": {
			"post": {
				"operationId": "CreateBranches",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IInstitution"
								}
							}
						}
					}
				},
				"tags": [
					"Institutions"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateBranch"
							}
						}
					}
				}
			}
		},
		"/api/institutions/all": {
			"get": {
				"operationId": "All",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IInstitutionDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Institutions"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/import/drugs": {
			"post": {
				"operationId": "UploadFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Import"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"file"
								]
							}
						}
					}
				}
			}
		},
		"/api/import/exams": {
			"post": {
				"operationId": "ImportExams",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Import"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"file"
								]
							}
						}
					}
				}
			}
		},
		"/api/import/insurance-prices/{type}": {
			"post": {
				"operationId": "ImportInsurancePrices",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Import"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"file"
								]
							}
						}
					}
				}
			}
		},
		"/api/import/institution-prices/{type}": {
			"post": {
				"operationId": "ImportInstitutionPrice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Import"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"file": {
										"type": "string",
										"format": "binary"
									}
								},
								"required": [
									"file"
								]
							}
						}
					}
				}
			}
		},
...
		"/api/exams": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaged_IExamResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Exams"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IExam"
								}
							}
						}
					}
				},
				"tags": [
					"Exams"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IExamRequest"
							}
						}
					}
				}
			}
		},
...
		"/api/drugs": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaged_IDrugResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Drugs"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IDrugDTO"
								}
							}
						}
					}
				},
				"tags": [
					"Drugs"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IDrugRequest"
							}
						}
					}
				}
			}
		},
		"/api/drugs/institution": {
			"get": {
				"operationId": "GetInstitutionDrugs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaged_IInstitutionDrugResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Drugs"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/drugs/institution/all": {
			"get": {
				"operationId": "InstitutionAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IInstitutionDrug"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Drugs"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}